requirements.txt
Flask
python-telegram-bot==20.7
requests